refactor code to generate subnet files
