+++ container-manager/src/libs/exec.js
@@ -1,7 +1,9 @@
 const exec = require("child_process").exec;
+const { execSync } = require("child_process");
 const path = require("path");
 const mountPath = path.join(__dirname, "../../mount/generated/");
 const config = require("./config");
+const fs = require('fs')
 
 module.exports = {
   startComposeProfile,
@@ -10,6 +12,7 @@ module.exports = {
   executeTest,
   startSubnetGradual,
   removeSubnet,
+  generate,
 };
 
 const streamExecOutput = (data) => {
@@ -130,3 +133,142 @@ async function execute(command, outputHandler, doneHandler) {
   return prom;
 }
 
+function generate(params){
+  genEnv = genGenEnv(params)
+  fs.writeFileSync(path.join(__dirname, "../../mount/generated/gen.env"), genEnv, (err) => { //write to mount
+    if (err) {
+      console.error(err);
+      exit();
+    }
+  });
+  return callExec(`cd ${__dirname}/../gen; node gen.js`);
+}
+
+function callExec(command) {
+  try {
+    const stdout = execSync(command, { timeout: 200000, encoding: 'utf-8'});
+    output = stdout.toString();
+    // console.log(output);
+    return [true, output]
+  } catch (error) {
+    console.log(error)
+    return [false, error.stdout]
+    throw Error(error.stdout);
+  }
+}
+
+function genGenEnv(input){
+  console.log(input)
+
+  let content_machine = ''
+  if (input["text-num-machine"] > 1){
+    content_machine += `\nMAIN_IP=${input["text-private-ip"]}`
+    if (input["text-public-ip"] != ''){
+      content_machine += `\nPUBLIC_IP=${input["text-public-ip"]}`  
+    }
+    if (input["text-num-machine"] != ''){
+      content_machine += `\nNUM_MACHINE=${input["text-num-machine"]}`
+    }
+  } else {
+    content_machine += `\nMAIN_IP=127.0.0.1`
+    content_machine += `\nNUM_MACHINE=1`
+  }
+
+  let parentnet = ''
+  switch (input.pnradio){
+    case 'pn-radio-testnet':
+      parentnet = 'testnet'
+      break
+    case 'pn-radio-devnet':
+      parentnet = 'devnet'
+      break
+    case 'pn-radio-mainnet':
+      parentnet = 'mainnet'
+      break
+  }
+
+  let relayer_mode = ''
+  switch (input.rmradio){
+    case 'rm-radio-full':
+      relayer_mode = 'full'
+      break
+    case 'rm-radio-lite':
+      relayer_mode = 'lite'
+      break
+  }
+
+  let content_custom_key = ''
+  if (input["grandmaster-pk"] != ''){
+    content_custom_key += `\nGRANDMASTER_PK=${input["grandmaster-pk"]}`
+  }
+
+  let subnet_keys=[]
+  let idx = 1
+  while ('subnet-key'+idx.toString() in input){
+    key = 'subnet-key'+idx.toString()
+    subnet_keys.push(input[key])
+    idx++
+  }
+  if (subnet_keys.length > 0){
+    key_string = subnet_keys.join(',')
+    content_custom_key += `\nSUBNETS_PK=${key_string}`
+  }
+
+  let content_version = ''
+  if (input["customversion-subnet"] != ''){
+    content_version += `\nVERSION_SUBNET=${input["customversion-subnet"]}`
+  }
+  if (input["customversion-bootnode"] != ''){
+    content_version += `\nVERSION_BOOTNODE=${input["customversion-bootnode"]}`
+  }
+  if (input["customversion-relayer"] != ''){
+    content_version += `\nVERSION_RELAYER=${input["customversion-relayer"]}`
+  }
+  if (input["customversion-stats"] != ''){
+    content_version += `\nVERSION_STATS=${input["customversion-stats"]}`
+  }
+  if (input["customversion-frontend"] != ''){
+    content_version += `\nVERSION_FRONTEND=${input["customversion-frontend"]}`
+  }
+  if (input["customversion-csc"] != ''){
+    content_version += `\nVERSION_CSC=${input["customversion-csc"]}`
+  }
+  if (input["customversion-zero"] != ''){
+    content_version += `\nVERSION_ZERO=${input["customversion-zero"]}`
+  }
+
+  let content_zero = ''
+  if (relayer_mode == 'full' && 'xdczero-checkbox' in input){
+    if (input["zmradio"] == 'zm-radio-one'){
+      content_zero += '\nXDC_ZERO=one-directional'
+    }
+    if (input["zmradio"] == 'zm-radio-bi'){
+      content_zero += '\nXDC_ZERO=bi-directional'
+      content_zero += `\nSUBNET_WALLET_PK=${input["subnet-wallet-pk"]}`
+      content_zero += `\nSUBNET_ZERO_WALLET_PK=${input["subnet-zero-wallet-pk"]}` 
+    }
+    content_zero += `\nPARENTNET_ZERO_WALLET_PK=${input["parentnet-zero-wallet-pk"]}`
+    if ('subswap-checkbox' in input){
+      content_zero += '\nSUBSWAP=true'
+    }
+  }
+
+  content=`
+NETWORK_NAME=${input["text-subnet-name"]}
+NUM_SUBNET=${input["text-num-subnet"]}
+PARENTNET=${parentnet}
+PARENTNET_WALLET_PK=${input["parentnet-wallet-pk"]}
+RELAYER_MODE=${relayer_mode}
+`
+  content+=content_machine
+  content+='\n'
+  content+=content_custom_key
+  content+='\n'
+  content+=content_version
+  content+='\n'
+  content+=content_zero
+
+  console.log(content)
+
+  return content
+}
